#  Remove ExceptionTypeCache