--- a/conda.recipe/meta.yaml
+++ b/conda.recipe/meta.yaml
@@ -1,6 +1,6 @@
 package:
   name: pythonnet
-  version: "1.0.5.16"
+  version: "1.0.5.17"
 
 build:
   skip: True  # [not win]