--- a/setup.py
+++ b/setup.py
@@ -485,7 +485,7 @@ if not os.path.exists(_get_interop_filename()):
 
 setup(
     name="pythonnet",
-    version="1.0.5.16",
+    version="1.0.5.17",
     description=".Net and Mono integration for Python",
     url='https://pythonnet.github.io/',
     license='MIT',