- Remove previously added `ExceptionTypeCache` which was causing issues and fix for issue
caused performance overhead.
- Bumping version to 1.0.5.17

Performance testing:
[MASTER C# IndicatorRibbonBenchmark] From PR https://github.com/QuantConnect/pythonnet/pull/24
132.27 seconds at 6k data points per second. Processing total of 782,223 data points.
130.07 seconds at 6k data points per second. Processing total of 782,223 data points.
129.88 seconds at 6k data points per second. Processing total of 782,223 data points.
[This PR C# IndicatorRibbonBenchmark]
127.63 seconds at 6k data points per second. Processing total of 782,223 data points.
127.83 seconds at 6k data points per second. Processing total of 782,223 data points.
127.44 seconds at 6k data points per second. Processing total of 782,223 data points.

**Results suggest the cache wasn't really giving a significant performance improvement.**